# steven/helix-prototypes/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "GateStatus": {
        "enum": [
          "blocked",
          "ready_for_review",
          "ready_for_signature",
          "ready_for_export",
          "exported"
        ],
        "title": "GateStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StudyListItem": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "release_status": {
            "$ref": "#/components/schemas/GateStatus"
          },
          "study_id": {
            "title": "Study Id",
            "type": "string"
          },
          "study_type_id": {
            "title": "Study Type Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_state": {
            "title": "Workflow State",
            "type": "string"
          }
        },
        "required": [
          "study_id",
          "study_type_id",
          "title",
          "workflow_state",
          "release_status",
          "label"
        ],
        "title": "StudyListItem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A synthetic pattern-testing API. It does not certify GLP, Part 11, SEND, eCTD, or FDA acceptance.",
    "title": "HELIX synthetic nonclinical workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/studies": {
      "get": {
        "operationId": "list_studies_api_v1_studies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudyListItem"
                  },
                  "title": "Response List Studies Api V1 Studies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Studies",
        "tags": [
          "studies"
        ]
      },
      "post": {
        "description": "Create a study from uploaded source files.\n\nRoute and protocol version are required because no column carries them\nand a governed fact is never inferred from data; the caller is recorded\nas their source.\n\nThe package is stored with no claims. Computing those is the section\nexecutor's job, so the release gate starts blocked and stays blocked\nuntil real validation runs.",
        "operationId": "create_study_from_upload_api_v1_studies_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_study_from_upload_api_v1_studies_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Study From Upload Api V1 Studies Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Study From Upload",
        "tags": [
          "studies"
        ]
      }
    },
    "/api/v1/studies/{study_id}/approvals": {
      "post": {
        "operationId": "record_approval_api_v1_studies__study_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Approval",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/studies/{study_id}/claims/{claim_id}/evidence": {
      "get": {
        "operationId": "get_evidence_api_v1_studies__study_id__claims__claim_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceChain"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/studies/{study_id}/data-validation-packages": {
      "post": {
        "operationId": "run_data_validation_api_v1_studies__study_id__data_validation_packages_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataValidationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataValidationExecution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Data Validation",
        "tags": [
          "data-validation"
        ]
      }
    },
    "/api/v1/studies/{study_id}/exports": {
      "post": {
        "operationId": "export_package_api_v1_studies__study_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Package",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/studies/{study_id}/exports/{artifact_id}": {
      "get": {
        "operationId": "download_artifact_api_v1_studies__study_id__exports__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/studies/{study_id}/final-study-approvals": {
      "post": {
        "operationId": "record_final_study_approval_api_v1_studies__study_id__final_study_approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalStudyApprovalCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Final Study Approval",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/studies/{study_id}/pinned-runs": {
      "post": {
        "operationId": "freeze_run_api_v1_studies__study_id__pinned_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeRunCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Run",
        "tags": [
          "run-plans"
        ]
      }
    },
...
  }
}